refactor: Move formulário de persona para Padrões de Campanha
Este commit refatora a implementação da funcionalidade de persona granular com base no seu feedback.

A implementação inicial colocou incorretamente o formulário detalhado na tela principal de 'Campanha'. Esta versão corrige isso movendo o formulário para o local correto, que é o modal de 'Padrões de Campanha'.

As principais alterações incluem:

- **Remoção do Formulário de `Campaign.jsx`**: Todo o código relacionado ao formulário de persona foi removido do componente principal da campanha.
- **Implementação em `CampaignStandardsModal.jsx`**: O formulário de persona completo e detalhado, com todos os seus campos, lógica condicional e dicas de ajuda, foi integrado na aba 'Persona' do modal de Padrões de Campanha.
- **Correção de Lógica**: A lógica de estado, carregamento e salvamento no modal foi ajustada para gerenciar o objeto de persona complexo.
- **Limpeza de Código**: Foram removidas variáveis não utilizadas e o código foi ajustado para passar no linter.

Esta implementação agora alinha a funcionalidade com os requisitos corretos, substituindo a antiga interface de edição de persona no modal pela nova versão granular.
